2019/CVE-2019-20933/requirements.txt
urllib3
requests
PyJWT
termcolor
